Rate Calculations
Added Fatemeh's new code with heart rate calculations and synchronized to the old code with few minor changes. New subject was added as it was forgotten in older versions (that is subject number 5). Added new code to calculate respiration rate from the thoracic impedance data. The calculations are implemented in the function getAvgRespRate.m. Also deleted some old files that are no longer in use