Implementation of QT interval calculations
Added the functions getTEnds and getQStart to find the ends of the T waves and the beginnings of the Q waves to calculate the QT intervals. The method used was the one decribed here: https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3201026/ . Next step is to actually calculate the intervals and then go into analysis.